requirements.txt
altair==4.2.2
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
cryptography==39.0.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fake-useragent==1.1.1
filelock==3.9.0
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.5
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodomex==3.16.0
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.0
requests==2.28.2
rich==13.3.1
selenium==4.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==3.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.17.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.12.1
